refactor(cleanup): move & rename files, remove comments, cov delete
