+++ src/Analysis/GWASWizard/GWASQuantitative.jsx
@@ -1,20 +1,19 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
-// import { useQuery, useMutation } from 'react-query';
 import {
   Steps, Button, Space, Popconfirm, Spin,
 } from 'antd';
 import CohortSelect from './shared/CohortSelect';
 import CovariateSelect from './shared/CovariateSelect';
 import { quantitativeSteps } from './shared/constants';
-import { useSourceFetch } from './wizard-endpoints/cohort-middleware-api';
+import { useSourceFetch } from './wizardEndpoints/cohortMiddlewareApi';
 import '../GWASUIApp/GWASUIApp.css';
 import AddCohortButton from './shared/AddCohortButton';
 import OutcomeSelectReview from './OutcomeSelectReview';
 import CustomDichotomousSelect from './shared/CustomDichotomousSelect';
 import WorkflowParameters from './shared/WorkflowParameters';
 import GWASFormSubmit from './shared/GWASFormSubmit';
-import TourButton from '../GWASUIApp/TourButton';
+import TourButton from './shared/TourButton';
 
 const { Step } = Steps;
 
@@ -38,6 +37,11 @@ const GWASQuantitative = ({ resetGWASType, refreshWorkflows }) => {
 
   const handleCovariateSelect = (cov) => {
     setSelectedCovariates(cov);
+  };
+
+  const handleCovariateDelete = (remainingCovariates) => {
+    const covariateMapping = remainingCovariates.map((conceptName) => selectedCovariates.find((concept) => concept.concept_name === conceptName));
+    setSelectedCovariates(covariateMapping);
   };
 
   const handleOutcomeSelect = (selectedOutcome) => {
@@ -93,20 +97,39 @@ const GWASQuantitative = ({ resetGWASType, refreshWorkflows }) => {
 
   const generateStep = () => {
     switch (current) {
-    case 0:
-      return (!loading && sourceId
-        ? (
+      case 0:
+        return (!loading && sourceId
+          ? (
+            <React.Fragment>
+              <div data-tour='quant-step-1-new-cohort'>
+                <AddCohortButton />
+              </div>
+              <React.Fragment>
+                <Space direction={'vertical'} align={'center'} style={{ width: '100%' }}>
+                  <h4 className='GWASUI-selectInstruction' data-tour='quant-step-1-cohort-selection'>In this step, you will determine the study population. To begin, select the cohort that you would like to define your study population with.</h4>
+                  <div className='GWASUI-mainTable'>
+                    <CohortSelect
+                      selectedCohort={selectedCohort}
+                      handleCohortSelect={handleCohortSelect}
+                      sourceId={sourceId}
+                      current={current}
+                    />
+                  </div>
+                </Space>
+                <TourButton stepInfo={stepInfo} />
+              </React.Fragment>
+            </React.Fragment>
+          ) : <Spin />);
+      case 1:
+        return (
           <React.Fragment>
-            <div data-tour='quant-step-1-new-cohort'>
-              <AddCohortButton />
-            </div>
             <React.Fragment>
               <Space direction={'vertical'} align={'center'} style={{ width: '100%' }}>
-                <h4 className='GWASUI-selectInstruction' data-tour='quant-step-1-cohort-selection'>In this step, you will determine the study population. To begin, select the cohort that you would like to define your study population with.</h4>
+                <h4 className='GWASUI-selectInstruction' data-tour='quant-step-2-choosing-variable'>In this step, you will select the harmonized variables for your study. Please select all variables you wish to use in your model, including both covariates and phenotype. (Note: population PCs are not included in this step)</h4>
                 <div className='GWASUI-mainTable'>
-                  <CohortSelect
-                    selectedCohort={selectedCohort}
-                    handleCohortSelect={handleCohortSelect}
+                  <CovariateSelect
+                    selectedCovariates={selectedCovariates}
+                    handleCovariateSelect={handleCovariateSelect}
                     sourceId={sourceId}
                     current={current}
                   />
@@ -115,101 +138,84 @@ const GWASQuantitative = ({ resetGWASType, refreshWorkflows }) => {
               <TourButton stepInfo={stepInfo} />
             </React.Fragment>
           </React.Fragment>
-        ) : <Spin />);
-    case 1:
-      return (
-        <React.Fragment>
+        );
+      case 2:
+        return (
           <React.Fragment>
-            <Space direction={'vertical'} align={'center'} style={{ width: '100%' }}>
-              <h4 className='GWASUI-selectInstruction' data-tour='quant-step-2-choosing-variable'>In this step, you will select the harmonized variables for your study. Please select all variables you wish to use in your model, including both covariates and phenotype. (Note: population PCs are not included in this step)</h4>
-              <div className='GWASUI-mainTable'>
-                <CovariateSelect
-                  selectedCovariates={selectedCovariates}
-                  handleCovariateSelect={handleCovariateSelect}
-                  sourceId={sourceId}
-                  current={current}
-                />
-              </div>
-            </Space>
-            <TourButton stepInfo={stepInfo} />
-          </React.Fragment>
-        </React.Fragment>
-      );
-    case 2:
-      return (
-        <React.Fragment>
-          <OutcomeSelectReview
-            cohortDefinitionId={selectedCohort.cohort_definition_id}
-            selectedCovariates={selectedCovariates}
-            outcome={outcome}
-            handleOutcomeSelect={handleOutcomeSelect}
-            sourceId={sourceId}
-            current={current}
-          />
-        </React.Fragment>
-      );
-    case 3:
-      return (
-        <React.Fragment>
-          <CustomDichotomousSelect
-            sourceId={sourceId}
-            handleCDAdd={handleCDAdd}
-            handleCDRemove={handleCDRemove}
-            selectedDichotomousCovariates={selectedDichotomousCovariates}
-            current={current}
-          />
-        </React.Fragment>
-      );
-    case 4:
-      return (
-        <React.Fragment>
-          <WorkflowParameters
-            quantitativeCohortDefinitionId={selectedCohort.cohort_definition_id}
-            selectedCovariates={selectedCovariates}
-            selectedDichotomousCovariates={selectedDichotomousCovariates}
-            sourceId={sourceId}
-            workflowType={'quantitative'}
-            numOfPC={numOfPC}
-            handleNumOfPC={handleNumOfPC}
-            mafThreshold={mafThreshold}
-            handleMaf={handleMaf}
-            imputationScore={imputationScore}
-            handleImputation={handleImputation}
-            selectedHare={selectedHare}
-            handleHareChange={handleHareChange}
-          />
-          <TourButton stepInfo={stepInfo} />
-        </React.Fragment>
-      );
-    case 5:
-      return (
-        <React.Fragment>
-          <h4 className='GWASUI-selectInstruction'>In this step, you may review the metadata selected for the study, give a name to the study, and submit the GWAS for analysis.</h4>
-          <h4 className='GWASUI-selectInstruction'>Upon submission you may review the status of the job in the ‘Submitted Job Status’ in this App above the enumerated steps</h4>
-          <div className='GWASUI-mainArea'>
-            <GWASFormSubmit
-              sourceId={sourceId}
-              numOfPC={numOfPC}
-              mafThreshold={mafThreshold}
-              imputationScore={imputationScore}
-              selectedHare={selectedHare}
-              selectedQuantitativeCohort={selectedCohort}
-              workflowType={'quantitative'}
+            <OutcomeSelectReview
+              cohortDefinitionId={selectedCohort.cohort_definition_id}
+              selectedCovariates={selectedCovariates}
               outcome={outcome}
-              // selectedCaseCohort={selectedCaseCohort}
-              // selectedControlCohort={selectedControlCohort}
+              handleOutcomeSelect={handleOutcomeSelect}
+              sourceId={sourceId}
+              current={current}
+            />
+          </React.Fragment>
+        );
+      case 3:
+        return (
+          <React.Fragment>
+            <CustomDichotomousSelect
+              sourceId={sourceId}
+              handleCDAdd={handleCDAdd}
+              handleCDRemove={handleCDRemove}
+              selectedDichotomousCovariates={selectedDichotomousCovariates}
+              current={current}
+            />
+          </React.Fragment>
+        );
+      case 4:
+        return (
+          <React.Fragment>
+            <WorkflowParameters
+              quantitativeCohortDefinitionId={selectedCohort.cohort_definition_id}
               selectedCovariates={selectedCovariates}
               selectedDichotomousCovariates={selectedDichotomousCovariates}
-              gwasName={gwasName}
-              handleGwasNameChange={handleGwasNameChange}
-              resetGWAS={resetQuantitative}
+              sourceId={sourceId}
+              workflowType={'quantitative'}
+              numOfPC={numOfPC}
+              handleNumOfPC={handleNumOfPC}
+              mafThreshold={mafThreshold}
+              handleMaf={handleMaf}
+              imputationScore={imputationScore}
+              handleImputation={handleImputation}
+              selectedHare={selectedHare}
+              handleHareChange={handleHareChange}
+              handleCovariateDelete={handleCovariateDelete}
+              outcomeId={outcome.concept_id}
             />
             <TourButton stepInfo={stepInfo} />
-          </div>
-        </React.Fragment>
-      );
-    default:
-      return <React.Fragment />;
+          </React.Fragment>
+        );
+      case 5:
+        return (
+          <React.Fragment>
+            <h4 className='GWASUI-selectInstruction'>In this step, you may review the metadata selected for the study, give a name to the study, and submit the GWAS for analysis.</h4>
+            <h4 className='GWASUI-selectInstruction'>Upon submission you may review the status of the job in the ‘Submitted Job Status’ in this App above the enumerated steps</h4>
+            <div className='GWASUI-mainArea'>
+              <GWASFormSubmit
+                sourceId={sourceId}
+                numOfPC={numOfPC}
+                mafThreshold={mafThreshold}
+                imputationScore={imputationScore}
+                selectedHare={selectedHare}
+                selectedQuantitativeCohort={selectedCohort}
+                workflowType={'quantitative'}
+                outcome={outcome}
+                // selectedCaseCohort={selectedCaseCohort}
+                // selectedControlCohort={selectedControlCohort}
+                selectedCovariates={selectedCovariates}
+                selectedDichotomousCovariates={selectedDichotomousCovariates}
+                gwasName={gwasName}
+                handleGwasNameChange={handleGwasNameChange}
+                resetGWAS={resetQuantitative}
+              />
+              <TourButton stepInfo={stepInfo} />
+            </div>
+          </React.Fragment>
+        );
+      default:
+        return <React.Fragment />;
     }
   };
