+++ src/Analysis/GWASWizard/shared/Covariates.jsx
@@ -3,9 +3,9 @@ import PropTypes from 'prop-types';
 import { useQuery } from 'react-query';
 import { Table, Spin } from 'antd';
 import { covariateTableConfig, covariateSelection } from './constants';
-import { fetchCovariates, queryConfig } from '../wizard-endpoints/cohort-middleware-api';
+import { fetchCovariates, queryConfig } from '../wizardEndpoints/cohortMiddlewareApi';
 import '../../GWASUIApp/GWASUIApp.css';
-import { useFetch, useFilter } from './form-hooks';
+import { useFetch, useFilter } from './formHooks';
 
 const Covariates = ({
   sourceId, searchTerm, selectedCovariates, handleCovariateSelect,
